Audit fixes (#17)
* update readme

* update encoding for settler and switch chainId to uint256

* add faqs

* change prettier line length to 120

* switch migrationHash back to migrationId

* reset client before settler test

* fix two bugs

* update tests and add a moduleMocker

* update test tokenId

+++ src/abis/MigrationParams.ts
@@ -5,8 +5,8 @@ export const MigrationParamsAbi = [
     components: [
       {
         name: 'chainId',
-        internalType: 'uint32',
-        type: 'uint32',
+        internalType: 'uint256',
+        type: 'uint256',
       },
       {
         name: 'settler',
